extract DynamicChunkSizer scheduler component
