+++ test/manual/chunked_prefill/test_scripted_special_case.py
@@ -885,8 +885,8 @@ class TestSpecialCaseDynamicChunkingPP1(ScriptedTestCase):
 
     @staticmethod
     def _script_dynamic_chunking_forced_off_on_pp1(t: ScriptedContext):
-        assert t.scheduler.enable_dynamic_chunking is False, (
-            "pp_size==1 must force enable_dynamic_chunking off even when the "
+        assert t.scheduler.dynamic_chunk_sizer is None, (
+            "pp_size==1 must leave dynamic chunking off even when the "
             "server arg is True (the 'and ps.pp_size > 1' conjunct)"
         )
         r = t.start_req(
